Add files via upload
Created Bank2 which implements BankInterface.  Attribute accounts of Bank2 will be stored as an ArrayList of Account objects.
In Bank2, defined all the methods promised in BankInterface; they will work with the accounts as an ArrayList, instead of an array, i.e. the data type of accounts will be ArrayList<Account> in my new class Bank2.
The body of your sortAccounts() method works like this: Collections.sort(accounts); where accounts is an ArrayList and Account implements Comparable and overrides compareTo method.  Both Collections and ArrayList are predefined classes in the Java Collections Framework (JCF); imported them from the java.util package.  The sort method is a static method of the predefined Collections class.

Created a JUnit class Bank2Test, that is similar to BankTest except it uses Bank2 instead of Bank.  Assuming your BankTest did not work with arrays (just worked with the methods as specified in the BankInterface), the only modification made, is to call the Bank2 constructor rather than the Bank constructor.  Note: If the data type of my bank object was Bank, then I'd need to change that to Bank2 (or even better: BankInterface); if BankInterface was the data type of your bank object, then no change will be needed there.